# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FD — Face Detection & Recognition Service
  description: |
    Production-ready REST API for real-time face detection, recognition,
    gender/age prediction from video streams (RTSP, YouTube, HTTP).
  version: 1.0.0
  contact:
    name: FD Team
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        collection_id:
          type: string
          format: uuid
        name:
          type: string
        metadata:
          type: object
        face_count:
          type: integer
        created_at:
          type: string
          format: date-time
    CreatePerson:
      type: object
      properties:
        collection_id:
          type: string
          format: uuid
        name:
          type: string
        metadata:
          type: object
      required: [collection_id, name]
    FaceEmbedding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        person_id:
          type: string
          format: uuid
        quality:
          type: number
          format: float
        source_key:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/persons:
    post:
      tags: [Persons]
      summary: Create a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePerson'
      responses:
        '201':
          description: Person created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
  /v1/persons/{id}:
    get:
      tags: [Persons]
      summary: Get person details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Person details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '404':
          description: Not found
  /v1/persons/{id}/faces:
    post:
      tags: [Persons]
      summary: Upload a face image to extract embedding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required: [image]
      responses:
        '201':
          description: Face added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceEmbedding'
        '422':
          description: No face detected in image
    get:
      tags: [Persons]
      summary: List face embeddings for a person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of face embeddings
...